refactor: AbstractController 패키지 위치 변경 및 sealed 키워드 제거
