+++ tomcat/src/main/java/com/techcourse/controller/RegisterController.java
@@ -6,7 +6,6 @@ import java.nio.file.Path;
 import java.util.LinkedHashMap;
 import java.util.Optional;
 import org.apache.catalina.Session;
-import org.apache.coyote.http11.AbstractController;
 import org.apache.coyote.http11.Http11Cookie;
 import org.apache.coyote.http11.Http11ResourceFinder;
 import org.apache.coyote.http11.request.HttpRequest;
