+++ tomcat/src/main/java/org/apache/coyote/http11/Controller.java
@@ -3,6 +3,6 @@ package org.apache.coyote.http11;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
 
-public sealed interface Controller permits AbstractController {
+public interface Controller {
     void service(HttpRequest request, HttpResponse response);
 }
